Refactoring to eliminate duplicate code
