+++ cdap-watchdog/src/test/java/co/cask/cdap/logging/appender/file/TestFileLogging.java
@@ -30,9 +30,9 @@ import co.cask.cdap.logging.appender.LoggingTester;
 import co.cask.cdap.logging.context.FlowletLoggingContext;
 import co.cask.cdap.logging.filter.Filter;
 import co.cask.cdap.logging.guice.LoggingModules;
+import co.cask.cdap.logging.read.FileLogReader;
 import co.cask.cdap.logging.read.LogEvent;
 import co.cask.cdap.logging.read.LogOffset;
-import co.cask.cdap.logging.read.StandaloneLogReader;
 import co.cask.tephra.TransactionManager;
 import co.cask.tephra.runtime.TransactionModules;
 import com.google.inject.Guice;
@@ -98,26 +98,24 @@ public class TestFileLogging {
   @Test
   public void testGetLogNext() throws Exception {
     LoggingContext loggingContext = new FlowletLoggingContext("TFL_NS_1", "APP_1", "FLOW_1", "", "RUN1", "INSTANCE1");
-    StandaloneLogReader logReader = injector.getInstance(StandaloneLogReader.class);
+    FileLogReader logReader = injector.getInstance(FileLogReader.class);
     LoggingTester tester = new LoggingTester();
     tester.testGetNext(logReader, loggingContext);
-    logReader.close();
   }
 
   @Test
   public void testGetLogPrev() throws Exception {
     LoggingContext loggingContext = new FlowletLoggingContext("TFL_NS_1", "APP_1", "FLOW_1", "", "RUN1", "INSTANCE1");
-    StandaloneLogReader logReader = injector.getInstance(StandaloneLogReader.class);
+    FileLogReader logReader = injector.getInstance(FileLogReader.class);
     LoggingTester tester = new LoggingTester();
     tester.testGetPrev(logReader, loggingContext);
-    logReader.close();
   }
 
   @Test
   public void testGetLog() throws Exception {
     // LogReader.getLog is tested in LogSaverTest for distributed mode
     LoggingContext loggingContext = new FlowletLoggingContext("TFL_NS_1", "APP_1", "FLOW_1", "", "RUN1", "INSTANCE1");
-    StandaloneLogReader logTail = injector.getInstance(StandaloneLogReader.class);
+    FileLogReader logTail = injector.getInstance(FileLogReader.class);
     LoggingTester.LogCallback logCallback1 = new LoggingTester.LogCallback();
     logTail.getLogPrev(loggingContext, LogOffset.LATEST_OFFSET, 60, Filter.EMPTY_FILTER,
                        logCallback1);
